+++ app/api/Files.scala
@@ -32,6 +32,7 @@ class Files @Inject()(
   collections: CollectionService,
   queries: MultimediaQueryService,
   tags: TagService,
+  sections: SectionService,
   comments: CommentService,
   extractions: ExtractionService,
   dtsrequests:ExtractionRequestsService,
@@ -206,6 +207,18 @@ class Files @Inject()(
         }
       }
     }
+
+  def getSections(id: UUID) = PermissionAction(Permission.ViewFile, Some(ResourceRef(ResourceRef.file, id))) { implicit request =>
+    implicit val user = request.user
+    files.get(id) match {
+      case Some(file) => {
+        val sectionList = sections.findByFileId(id)
+        val sectionIds = sectionList.map { section => section.id }
+        Ok(toJson(sectionIds))
+      }
+      case None => NotFound(toJson("The requested file does not exist"))
+    }
+  }
 
   def getMetadataDefinitions(id: UUID, space: Option[String]) = PermissionAction(Permission.AddMetadata, Some(ResourceRef(ResourceRef.file, id))) { implicit request =>
     implicit val user = request.user